Fix print statements
