+++ src/cocoa/repo.py
@@ -11,7 +11,7 @@ def is_git_repo(repo_path):
     return git.Repo(repo_path).git_dir is not None
 
 
-def get_remote_branches_info(repo_path):
+def get_remote_branches_info(repo_path, display=True):
     """
     This function reutrns the branch information from the
     remote repository.
@@ -30,11 +30,12 @@ def get_remote_branches_info(repo_path):
         num_ahead, num_behind = commits_diff.split("\t")
         branch_info.append([branch.name, num_ahead, num_behind])
 
-    for branch, behind, ahead in branch_info:
-        print(f"Branch: {branch}")
-        print(f"Commits behind main: {behind}")
-        print(f"Commits ahead of main: {ahead}")
-        print()
+    if display:
+        for branch, behind, ahead in branch_info:
+            print(f"Branch: {branch}")
+            print(f"Commits behind main: {behind}")
+            print(f"Commits ahead of main: {ahead}")
+            print()
 
     return branch_info
 
@@ -67,3 +68,30 @@ def get_current_branch(repo_path):
     except Exception as e:
         print(f"An unexpected error occurred: {e}")
     return None
+
+
+def check_branch_names(repo_path):
+    """
+    Check for branches other than 'main' and 'dev' in the repository.
+
+    Args:
+        repo_path (str): The path to the repository.
+
+    Returns:
+        list: A list of warnings regarding branch names.
+    """
+    branches_info = get_remote_branches_info(repo_path, False)
+    warnings = []
+
+    for branch_info in branches_info:
+        # Assuming branch_info is a list where the first item is the branch name
+        # Extracts branch name after 'origin/'
+        branch_name = branch_info[0].split("/")[-1]
+        if branch_name not in ["main", "dev"]:
+            warnings.append(
+                f"Warning: Found non-standard branch '{branch_name}' in repository."
+            )
+
+    for warning in warnings:
+        print(warning)
+    return warnings
